+++ gameResult.js
@@ -1,5 +1,4 @@
-import {player1} from './player.js';
-import {player2} from './player.js';
+import Player from './player.js';
 
 import {playerWin} from './utils.js';
 import {createElement} from './utils.js';
@@ -8,6 +7,20 @@ import {getRandom} from './utils.js';
 import * as variables from './variables.js';
 
 const {$arenas, $formFight, $chat, logs} = variables;
+
+const player1 = new Player({
+	player: 1,
+	name: 'Scorpion',
+	hp: 100,
+	img: 'http://reactmarathon-api.herokuapp.com/assets/scorpion.gif',
+});
+
+const player2 = new Player({
+	player: 2,
+	name: 'Kitana',
+	hp: 100,
+	img: 'http://reactmarathon-api.herokuapp.com/assets/kitana.gif',
+});
 
 
 function createReloadButton(){
